# NinjaClub/docs/swagger.yaml
openapi: 3.0.2
info:
    title: Ninja Club
    version: 1.0.0
    description: Ninja Club API
paths:
    /members:
        get:
            requestBody:
                content:
                    application/json: {}
                required: true
            tags:
                - members
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                List of members:
                                    value:
                                        - John Doe
                                        - Jane Smith
                                        - Peter Jones
                    description: Return a list of members
            summary: Get list of members from api
            description: Returns a list of all members in the API to the requesting user
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/member'
                        examples:
                            example of a member:
                                value:
                                    FirstName: Jane
                                    LastName: Smith
                                    Adress: Paju 5 Leht 3
                                    Phonenumber: '5557772222'
                                    Email: jane.smith@example.com
                                    RegistrationDate: 12.06.2005
                                    Active: false
                                    Level: 4
                required: true
            tags:
                - members
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                member that is added:
                                    value: |-
                                        [
                                            [
                                            "MemberID": 2,
                                            "FirstName": "Jane",
                                            "LastName": "Smith",
                                            "Adress": "Paju 5 Leht 3",
                                            "Phonenumber": "5557772222",
                                            "Email": "jane.smith@example.com",
                                            "RegistrationDate": "12.06.2005",
                                            "Active": false,
                                            "Level": 4
                                            ]
                                        ]
                    description: Return code 201 to user if creation of new member was successful.
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                error:
                                    value:
                                        -
                                            error: One or more parameters missing.
                    description: 'Return status 400 to the user, if one or multiple parameters are missing.'
            summary: Add new member
            description: Add new member into the array.
    '/members/{id}':
        get:
            tags:
                - members
            parameters:
                -
                    name: id
                    description: Id for Member the user has reported
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                EXAMPLE OF MEMBER:
                                    value:
                                        MemberID: 1
                                        FirstName: John2
                                        LastName: Doe
                                        Adress: Kuusk 1 Oks 2
                                        Phonenumber: '5556661111'
                                        Email: john.doe@example.com
                                        RegistrationDate: 01.11.2000
                                        Active: true
                                        Level: 6
                    description: Returns details of Member to the user
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Invalid member ID:
                                    value:
                                        -
                                            error: Invalid Id entered.
                    description: Return status code 400 to the user if invalid ID was entered.
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Member /w+ was not found:
                                    value:
                                        -
                                            error: Member\w+ was not found in member array.
                    description: Return error 404 to the user if member with this ID is not found.
            operationId: id
            summary: Get full details for a member
            description: |
                Returns full details of a given id member
        parameters:
            -
                examples:
                    Example if ID:
                        value: '1 '
                name: id
                description: This is id of member the user has requested.
                schema:
                    type: integer
                in: path
                required: true
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        member:
            title: Root Type for member
            description: example of submittable member
            type: object
            properties:
                FirstName:
                    type: string
                LastName:
                    type: string
                Adress:
                    type: string
                Phonenumber:
                    type: string
                Email:
                    type: string
                RegistrationDate:
                    type: string
                Active:
                    type: boolean
                Level:
                    format: int32
                    type: integer
            example:
                FirstName: Jane
                LastName: Smith
                Adress: Paju 5 Leht 3
                Phonenumber: '5557772222'
                Email: jane.smith@example.com
                RegistrationDate: 12.06.2005
                Active: false
                Level: 4